+++ admin/admin-api/src/main/java/dev/donggi/core/api/core/member/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package dev.donggi.core.api.core.exception;
+package dev.donggi.core.api.core.member.exception;
 
 import dev.donggi.core.api.common.exception.ErrorCodeAndMessage;
 import dev.donggi.core.api.common.exception.GolrabaException;
