+++ admin/admin-api/src/main/java/dev/donggi/core/api/web/AdminQuestionRestController.java
@@ -1,11 +1,11 @@
 package dev.donggi.core.api.web;
 
 import dev.donggi.core.api.common.response.ApiResponse;
-import dev.donggi.core.api.core.domain.AuthenticatedMember;
+import dev.donggi.core.api.core.member.domain.AuthenticatedMember;
 import dev.donggi.core.api.core.question.application.QuestionService;
 import dev.donggi.core.api.core.question.application.dto.QuestionDto;
 import dev.donggi.core.api.web.comment.dto.NoOffsetPageCommand;
-import dev.donggi.core.api.web.dto.LoginMember;
+import dev.donggi.core.api.web.member.dto.LoginMember;
 import dev.donggi.core.api.core.question.application.dto.QuestionPaginationDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/admin")
-public class AdminRestController {
+public class AdminQuestionRestController {
 
     private final QuestionService questionService;
