+++ admin/admin-api/src/main/java/dev/donggi/core/api/web/member/MemberRestController.java
@@ -1,11 +1,11 @@
-package dev.donggi.core.api.web;
+package dev.donggi.core.api.web.member;
 
-import dev.donggi.core.api.core.auth.AuthService;
-import dev.donggi.core.api.core.MemberService;
-import dev.donggi.core.api.core.dto.AccessAndRefreshTokenResponse;
-import dev.donggi.core.api.core.dto.LoginResponse;
-import dev.donggi.core.api.core.dto.MemberResponse;
-import dev.donggi.core.api.web.dto.LoginCommand;
+import dev.donggi.core.api.core.member.auth.AuthService;
+import dev.donggi.core.api.core.member.MemberService;
+import dev.donggi.core.api.core.member.dto.AccessAndRefreshTokenResponse;
+import dev.donggi.core.api.core.member.dto.LoginResponse;
+import dev.donggi.core.api.core.member.dto.MemberResponse;
+import dev.donggi.core.api.web.member.dto.LoginCommand;
 import dev.donggi.core.api.common.response.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
